+++ tests/SPIDAuthBaseTestCase.php
@@ -41,6 +41,8 @@ class SPIDAuthBaseTestCase extends TestCase
     protected function tearDown(): void
     {
         m::close();
+
+        parent::tearDown();
     }
 
     protected function getPackageProviders($app)